requirements.txt
numpy==1.15.4
pyspark==2.3.1
deap==1.2.2
scikit-learn==0.20.0
locustio==0.14.5
ruamel.yaml==0.16.10